Add a polynomial-based erf() calculation.
This implementation agrees with the Rust implementation to about 6 decimal places.
The Rust one is significantly more accurate (at least compared to LibreOffice),
but the polynomial is easier to get right in JS and the error margin is slightly
under the maximum limit of what would be acceptable for IPF Points.